+++ FcmSharp/FcmSharp/Settings/IFcmClientSettings.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Philipp Wagner. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using FcmSharp.BackOff;
+
 namespace FcmSharp.Settings
 {
     public interface IFcmClientSettings
@@ -8,5 +10,7 @@ namespace FcmSharp.Settings
         string Project { get; }
         
         string Credentials { get; }
+
+        ExponentialBackOffSettings ExponentialBackOffSettings { get; }
     }
 }